# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Dozhim API
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
  schemas:
    JobAccepted:
      type: object
      required: [job_id, status]
      properties:
        job_id:
          type: string
        status:
          type: string
          enum: [queued]
    Error:
      type: object
      required: [code, message, details, retryable, correlation_id]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        retryable:
          type: boolean
        correlation_id:
          type: string
...
paths:
...
  /campaigns/personalized/upload:
    post:
      summary: Upload excel/csv import for campaign
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, dry_run]
              properties:
                file:
                  type: string
                  format: binary
                dry_run:
                  type: boolean
      responses:
        "202":
          description: Accepted
          headers:
            Location:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobAccepted"
        "422":
          description: IMPORT_VALIDATION_FAILED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /jobs:
    get:
      summary: List jobs
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, default: 50 }
      responses:
        "200":
          description: OK
  /jobs/{jobId}:
    get:
      summary: Get job status
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
  /jobs/{jobId}/cancel:
    post:
      summary: Cancel running job
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - name: jobId
          in: path
          required: true
          schema: { type: string }
      responses:
        "202":
          description: Accepted
          headers:
            Location:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobAccepted"